Make extract_edge.py run on sample dataset by default
To test functionality, I wanted to run extract_edge.py on the dataset provided in folder assets/040119_PtK1_S01_01_phase_ROI2 however the path in user_params.py is not set to a path that exists. This is a minor tweak, since most people will be inputting their own datasets, but thought it would be nice for extract_edge.py to work out of the box.